manage_server/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
pymysql==1.1.1
python-dotenv== 0.21.0
PyJWT==2.8.0
Flask-Mail==0.10.0
openpyxl==3.1.2